Commit 5- Spiro Improvements
Changed set up of TheDragon class. When class is instantiated all you
need to pass is width and height of the JFrame. If we want to resize the
window we can sort that out a bit later. To make a Spiro object just
instantiate the class (it's nested atmo but if that is a problem you can
just take it out). The Spiro class takes some important constructors
that you NEED to pass into it. If you have other variables you want to
pass in let me know and I'll fix the script. Other than that the only
thing you need to know is that the script only supports 1 Spiro at the
moment, and if you want the buffered image you just have to call
getImage which will return a tasty image for you to draw to the screen.
I've introduced delta time into it so the animation should be silky
smooth. GetColour() probably doesn't work since I couldn't test it, but
pls don't mess around with my logic :P