Commit 7
Okay so now you need to do the following:

1. Create a SpiroObject with the necessary variables.
2. Initialise TheDragon instance passing the frame width and height, a
graphics instance you are using to create a buffered image, an empty
buffered image, and the SpiroObject.
3. Call getImage() every frame, and draw the image you get from it into
the JFrame.

Notes: Don't put any JFrame logic into TheDragon. It no longer extends
JFrame and should stay that way. getColour() is deliberately commented
out at the moment. Variable 't' in the parametric equations is a
variable between 0 and 2 pi- if you look at the maths you can see I am
dividing it by 50,000, if the program is slow feel free to reduce this.